Add a feature flag that controls PIN caching.
While SecretVec prevents accidentally logging or printing the PIN,
caching the PIN in RAM in a long-running application would presumably
allow it to be read from memory, or even from disk if the program gets
swapped out.

This commit adds a feature (enabled by default) that controls PIN
caching. If the `cache-pin` feature is disabled, the PIN is not cached
in memory and some YubiKey methods become unavailable.