PIV: Support AES management keys
This commit adds support for setting and getting the AES management
keys, these are available in firmwars 5.4 and later, and are now
the default in firmwares 5.7.

The key is handled via being generic on a limit number of allowed alogrithms,
using implementations of those from rust-crypto crates.

Right now support in PIV MGM keys is for:

* TripleDes (`0x03`) - The key type originally used
* AES128 (`0x08`) - The new key type using a 128 bit key
* AES192 (`0x0A`) - The new key type using a 192 bit key, this
    also doubles as the algorithm for firmwares 5.7 and later,
    where the default key is the same as the original TripleDes key.
* AES256 (`0x0C`) - The new key type using a 256 bit key

Suitable type aliases are provided for each of these key types.

The rationale here for exposing the key as a generic type parameter is
to largely use the original logic, but avoid scattered enums and provide
the end user with some degree of control over the key types at compile
time (it should, for instance be relatively easy make 3Des keys
uncompileable).

See: https://docs.yubico.com/yesdk/users-manual/application-piv/apdu/auth-mgmt.html